Move instanceof and exceptions removal back to goto-programs
