+++ src/goto-programs/remove_exceptions.cpp
