+++ src/goto-programs/remove_exceptions.h
