+++ src/goto-programs/remove_instanceof.cpp
