+++ src/goto-programs/remove_instanceof.h
