+++ src/java_bytecode/Makefile
@@ -31,8 +31,6 @@ SRC = bytecode_info.cpp \
       java_types.cpp \
       java_utils.cpp \
       mz_zip_archive.cpp \
-      remove_exceptions.cpp \
-      remove_instanceof.cpp \
       remove_java_new.cpp \
       replace_java_nondet.cpp \
       select_pointer_type.cpp \
