+++ src/java_bytecode/java_entry_point.cpp
@@ -8,15 +8,30 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "java_entry_point.h"
 
-#include <util/config.h>
-#include <util/string_constant.h>
-#include <util/suffix.h>
+#include <algorithm>
+#include <set>
+#include <unordered_set>
 
-#include <goto-programs/goto_functions.h>
+#include <goto-programs/remove_exceptions.h>
 
 #include <linking/static_lifetime_init.h>
 
+#include <util/arith_tools.h>
+#include <util/c_types.h>
+#include <util/config.h>
+#include <util/cprover_prefix.h>
+#include <util/message.h>
+#include <util/namespace.h>
+#include <util/pointer_offset_size.h>
+#include <util/prefix.h>
+#include <util/std_code.h>
+#include <util/std_expr.h>
+#include <util/std_types.h>
+#include <util/string_constant.h>
+#include <util/suffix.h>
+
 #include "java_object_factory.h"
+#include "java_types.h"
 #include "java_utils.h"
 
 static void create_initialize(symbol_table_baset &symbol_table)
